requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
cryptography==44.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django_notifications==0.1.dev0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.2.3
